TT | Splitting the Netflux app declaration into a full JSON and JSON+JSX declarations; Consolidated the library code into one module until the project is published to NPM.
